Add jumbotron to single template
Doing this required making edits in lots of places; the container opened
in the header has been moved into the necessary places for the
different templates since a jumbotron inside a container will not be
full-width.

Page titles have moved to the jumbotron.

This commit adds front-matter for content: `description`, which is the
subtitle in the jumbotron.

Signed-off-by: Ethan Madison <ethan@ethanmad.com>